requirement-test.txt
flake8>=3.3.0
tox>=2.7.0

pylint>=1.7.1
pylint-django>=0.7.2

pytest>=3.1.1
pytest-cov>=2.5.1
pytest-django>=3.1.2
mixer>=5.6.6
